requirements.txt
Flask
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
SpeechRecognition
gTTS
pydub
python-dotenv
